backend/requirements.txt
Django
djangorestframework
psycopg2-binary
python-decouple
dj-database-url
pytest
pytest-django
black
django-filter
pylint